chore: Dialog files moving
