+++ src/components/favorites/FavoritesFriendTab.vue
@@ -78,7 +78,7 @@
 
 <script>
     import FavoritesFriendItem from './FavoritesFriendItem.vue';
-    import FriendExportDialog from '../../views/dialogs/FriendExportDialog.vue';
+    import FriendExportDialog from '../../views/dialogs/favorites/FriendExportDialog.vue';
     import { favoriteRequest } from '../../classes/request';
 
     export default {
