+++ src/views/dialogs/favorites/FriendExportDialog.vue
